Struct protocol::utils::InitialMessage

pub struct InitialMessage {
    pub identity_key: PublicKey,
    pub ephemeral_key: PublicKey,
    pub prekey_hash: Sha256Hash,
    pub one_time_key_hash: Option<Sha256Hash>,
    pub challenge: Challenge,
    pub associated_data: AssociatedData,
}

A message sent by the initiator in the X3DH key exchange protocol.

Fields
identity_key: PublicKey

The initiatorโs identity public key.

ephemeral_key: PublicKey

The initiatorโs ephemeral public key.

prekey_hash: Sha256Hash

The SHA-256 hash of the responderโs signed pre-key.

one_time_key_hash: Option<Sha256Hash>

Optional SHA-256 hash of the responderโs one-time pre-key.

challenge: Challenge

A challenge generated by the initiator for authentication.

associated_data: AssociatedData

Associated identity key data for both parties.

Implementations

impl InitialMessage

pub(crate) const BASE_SIZE: usize = 208usize

The base byte size without an optional one-time prekey hash.

pub(crate) const SIZE_WITH_OTPK: usize = 240usize

The total byte size of the message when the one-time prekey hash is included.

pub fn get_associated_data(&self) -> AssociatedData

Returns a clone of the AssociatedData from the current message.

Returns

AssociatedData - The associated data contained in the message.

pub fn to_bytes(self) -> Vec<u8>

Converts the current InitialMessage into bytes.

Returns

Vec<u8> - A vector of bytes derived from the current InitialMessage.

pub fn to_base64(self) -> String

Converts the current InitialMessage into a base64-encoded string.

Returns

String - The base64-encoded string of the current InitialMessage.

pub fn size(&self) -> usize

Calculates the size of the current InitialMessage.

Returns

usize - The size of the current InitialMessage:

Self::BASE_SIZE - If there is no one-time prekey hash.

Self::SIZE_WITH_OTPK - If there is a one-time prekey hash.