Struct protocol::utils::PreKeyBundle

pub struct PreKeyBundle {
    pub verifying_key: VerifyingKey,
    pub ik: PublicKey,
    pub spk: PublicKey,
    pub sig: Signature,
    pub otpk: Vec<PublicKey>,
}

A PreKeyBundle contains the public keys and signature published by a recipient, used by an initiator to establish a shared secret using the X3DH key agreement protocol.

Fields
verifying_key: VerifyingKey

The recipient’s identity signing key (Ed25519), used to verify sig. For more information, see VerifyingKey.

ik: PublicKey

The recipient’s identity public key. For more information, see PublicKey.

spk: PublicKey

The recipient’s signed public pre-key. For more information, see PublicKey.

sig: Signature

A signature of the spk, signed by the identity signing key. For more information, see Signature.

otpk: Vec<PublicKey>

One or more ephemeral one-time pre-keys, X25519 public keys. If present, the initiator may use one to enhance forward secrecy. For more information, see PublicKey.

Implementations

impl PreKeyBundle

pub(crate) const BASE_SIZE: usize = 160usize

The total byte size of the pre-key bundle, which includes three Curve25519 public keys and one signature. This constant is used to verify the expected size of a PreKeyBundle.

pub fn new(ik: &PrivateKey, spk: PublicKey) -> Self

Generates a new pre-key bundle.

This method does not generate one-time pre-keys.
For that functionality, see PreKeyBundle::new_with_otpk.

Arguments

ik - The recipient’s identity key.

spk - The recipient’s signed pre-key.

Returns

PreKeyBundle - A PreKeyBundle struct.

pub fn new_with_otpk( ik: &PrivateKey, spk: PublicKey, otpk: Vec<PublicKey>, ) -> Self

Generates a new pre-key bundle, including one-time pre-keys.

For a version that excludes one-time pre-keys, see PreKeyBundle::new.

Arguments

ik - The recipient’s identity key.

spk - The recipient’s signed pre-key.

Returns

PreKeyBundle - A PreKeyBundle struct.

pub fn add_otpk(&mut self, otpk: PublicKey)

Adds a one-time pre-key

Arguments

otpk - The one-time pre-key to be added.

pub fn size(&self) -> usize

Calculates the size of the pre-key bundle.

Returns

usize - The number of elements in the pre-key bundle.

pub fn to_bytes(&self) -> Vec<u8>

Converts each element of the pre-key bundle into bytes.

Returns

Vec<u8> - A vector containing the byte representation of each element in the pre-key bundle.

pub fn to_base64(self) -> String

Calculates the base64 of the pre-key bundle.

Returns

String - The base64-encoded string of the pre-key bundle.